+++ src/client/cypress/e2e/editor/groundwaterLevelMeasurement.cy.js
@@ -5,6 +5,7 @@ import {
   createGroundwaterLevelMeasurement,
   handlePrompt,
   loginAsAdmin,
+  selectLanguage,
   startBoreholeEditing,
 } from "../helpers/testHelpers";
 import { evaluateDisplayValue, setInput, setSelect } from "../helpers/formHelpers";
@@ -36,18 +37,14 @@ describe("Tests for the groundwater level measurement editor.", () => {
       force: true,
     });
 
-    // switch to german
-    cy.contains("span", "DE").click({ force: true });
-    cy.wait(1000);
+    selectLanguage("de");
 
     // create groundwater level measurement
     addItem("addGroundwaterLevelMeasurement");
     cy.wait("@casing_GET");
 
     setSelect("kindId", 3);
-    setSelect("reliabilityId", 2);
     setSelect("casingId", 2);
-    setInput("startTime", "2012-11-14T12:06");
     setInput("levelM", "789.12");
     setInput("levelMasl", "5.4567");
 
@@ -57,7 +54,6 @@ describe("Tests for the groundwater level measurement editor.", () => {
     evaluateDisplayValue("gwlm_kind", "Manometer");
     evaluateDisplayValue("gwlm_levelm", "789.12");
     evaluateDisplayValue("gwlm_levelmasl", "5.4567");
-    evaluateDisplayValue("reliability", "fraglich");
 
     // edit groundwater level measurement
     startEditing();