+++ src/client/src/commons/menu/detailView/detailSideNav.jsx
@@ -196,23 +196,6 @@ const DetailSideNav = ({ borehole, history, match }) => {
                 </List.Content>
               </ChildListItem>
               <ChildListItem
-                active={location.pathname === `/${id}/hydrogeology/hydrotest`}
-                onClick={() => {
-                  history.push(`/${id}/hydrogeology/hydrotest`);
-                }}>
-                <List.Content>
-                  <List.Header as="h3" data-cy="hydrotest-menu-item">
-                    <Typography>
-                      <TranslationText firstUpperCase id="hydrotest" />
-                    </Typography>
-                  </List.Header>
-                </List.Content>
-              </ChildListItem>
-            </>
-          )}
-          {hydrogeologyIsVisible && (
-            <>
-              <ChildListItem
                 active={location.pathname === `/${id}/hydrogeology/groundwaterlevelmeasurement`}
                 onClick={() => {
                   history.push(`/${id}/hydrogeology/groundwaterlevelmeasurement`);
@@ -234,6 +217,19 @@ const DetailSideNav = ({ borehole, history, match }) => {
                   <List.Header as="h3" data-cy="fieldmeasurement-menu-item">
                     <Typography>
                       <TranslationText firstUpperCase id="fieldMeasurement" />
+                    </Typography>
+                  </List.Header>
+                </List.Content>
+              </ChildListItem>
+              <ChildListItem
+                active={location.pathname === `/${id}/hydrogeology/hydrotest`}
+                onClick={() => {
+                  history.push(`/${id}/hydrogeology/hydrotest`);
+                }}>
+                <List.Content>
+                  <List.Header as="h3" data-cy="hydrotest-menu-item">
+                    <Typography>
+                      <TranslationText firstUpperCase id="hydrotest" />
                     </Typography>
                   </List.Header>
                 </List.Content>